requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]